Refactor: introduce import_utils.py
To collect format-independent functions for model import.
