+++ python/amici/ode_export.py
@@ -45,7 +45,7 @@ from . import (
 )
 from .logging import get_logger, log_execution_time, set_log_level
 from .constants import SymbolId
-
+from .import_utils import smart_subs_dict
 
 # Template for model simulation main.cpp file
 CXX_MAIN_TEMPLATE_FILE = os.path.join(amiciSrcPath, 'main.template.cpp')
@@ -3528,47 +3528,6 @@ def cast_to_sym(value: Union[SupportsFloat, sp.Expr, BooleanAtom],
                         f"{type(value)}")
 
     return value
-
-
-SymbolDef = Dict[sp.Symbol, Union[Dict[str, sp.Expr], sp.Expr]]
-
-
-def smart_subs_dict(sym: sp.Expr,
-                    subs: SymbolDef,
-                    field: Optional[str] = None,
-                    reverse: bool = True) -> sp.Expr:
-    """
-    Subsitutes expressions completely flattening them out. Requires
-    sorting of expressions with toposort.
-
-    :param sym:
-        Symbolic expression in which expressions will be substituted
-
-    :param subs:
-        Substitutions
-
-    :param field:
-        Field of substitution expressions in subs.values(), if applicable
-
-    :param reverse:
-        Whether ordering in subs should be reversed. Note that substitution
-        requires the reverse order of what is required for evaluation.
-
-    :return:
-        Substituted symbolic expression
-    """
-    s = [
-        (eid, expr[field] if field is not None else expr)
-        for eid, expr in subs.items()
-    ]
-    if reverse:
-        s.reverse()
-    for substitution in s:
-        # note that substitution may change free symbols, so we have to do
-        # this recursively
-        if substitution[0] in sym.free_symbols:
-            sym = sym.subs(*substitution)
-    return sym
 
 
 @contextlib.contextmanager
